+++ packages/devtools_app/lib/src/standalone_ui/ide_shared/property_editor/utils/_utils_desktop.dart
@@ -1,0 +1,11 @@
+// Copyright 2025 The Flutter Authors
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
+
+void addBlurListener() {
+  // No-op for desktop platforms.
+}
+
+void removeBlurListener() {
+  // No-op for desktop platforms.
+}