+++ packages/devtools_app/test/shared/eval_integration_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:devtools_app/src/shared/globals.dart';
 import 'package:devtools_app_shared/service.dart';
+import 'package:devtools_app_shared/utils.dart';
 import 'package:devtools_test/helpers.dart';
 import 'package:flutter_test/flutter_test.dart';