+++ packages/devtools_app_shared/CHANGELOG.md
@@ -3,9 +3,12 @@ Copyright 2025 The Flutter Authors
 Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 -->
-## 0.3.2 (not released)
+## 0.4.0 (not released)
 * Add `DisposableController.init` method and `DisposableController.disposed`
 getter.
+* Move the `Disposable` class from the 
+`package:devtools_app_shared/service.dart` library to the 
+`package:devtools_app_shared/utils.dart` library.
 
 ## 0.3.1
 * Bump `vm_service` dependency to `>=13.0.0 <16.0.0`.